Add RPi support
Add RPi host support, with a generic setup aiming to allow multiple RPis
to share the same config, with just additional 'services' (yet to be
written) as the application software running on each system.
This layout aims to make it easy to change which host different services
are running on at any given time.